Changes to shape and Main file

+++ src/Shape.java
@@ -1,6 +1,6 @@
 public abstract class Shape {
     public abstract double area();
-    public int compareTo(Shape f){
-        return (int)(this.area()-f.area());
+    public double compareTo(Shape f){
+        return (double)(this.area()-f.area());
     }
 }
